Move mc_info scripting class into own folder
