+++ code/scripting/lua.cpp
@@ -80,6 +80,7 @@
 #include "scripting/api/texture.h"
 #include "scripting/api/texturemap.h"
 #include "scripting/api/weaponclass.h"
+#include "scripting/api/mc_info.h"
 
 using namespace scripting;
 using namespace scripting::api;
@@ -244,173 +245,6 @@ flag_def_list plr_commands[] = {
 };
 
 int num_plr_commands = sizeof(plr_commands)/sizeof(flag_def_list);
-
-class mc_info_h
-{
-protected:
-	mc_info* info;
-public:
-	mc_info_h(mc_info* val) : info(val) {}
-
-	mc_info_h() : info(NULL) {}
-
-	mc_info *Get()
-	{
-		return info;
-	}
-
-	void deleteInfo()
-	{
-		if (!this->IsValid())
-			return;
-
-		delete info;
-
-		info = NULL;
-	}
-
-	bool IsValid()
-	{
-		return info != NULL;
-	}
-};
-
-//**********HANDLE: Collision info
-ade_obj<mc_info_h> l_ColInfo("collision info", "Information about a collision");
-
-ADE_FUNC(__gc, l_ColInfo, NULL, "Removes the allocated reference of this handle", NULL, NULL)
-{
-	mc_info_h* info;
-
-	if(!ade_get_args(L, "o", l_ColInfo.GetPtr(&info)))
-		return ADE_RETURN_NIL;
-
-	if (info->IsValid())
-		info->deleteInfo();
-
-	return ADE_RETURN_NIL;
-}
-
-ADE_VIRTVAR(Model, l_ColInfo, "model", "The model this collision info is about", "model", "The model")
-{	
-	mc_info_h* info;
-	model_h * mh = nullptr;
-
-	if(!ade_get_args(L, "o|o", l_ColInfo.GetPtr(&info), l_Model.GetPtr(&mh)))
-		return ade_set_error(L, "o", l_Model.Set(model_h()));
-
-	if (!info->IsValid())
-		return ade_set_error(L, "o", l_Model.Set(model_h()));
-
-	mc_info *collide = info->Get();
-
-	int modelNum = collide->model_num;
-
-	if (ADE_SETTING_VAR && mh)
-	{
-		if (mh->IsValid())
-		{
-			collide->model_num = mh->GetID();
-		}
-	}
-
-	return ade_set_args(L, "o", l_Model.Set(model_h(modelNum)));
-}
-
-ADE_FUNC(getCollisionDistance, l_ColInfo, NULL, "The distance to the closest collision point", "number", "distance or -1 on error")
-{
-	mc_info_h* info;
-
-	if(!ade_get_args(L, "o", l_ColInfo.GetPtr(&info)))
-		return ade_set_error(L, "f", -1.0f);
-
-	if (!info->IsValid())
-		return ade_set_error(L, "f", -1.0f);
-
-	mc_info *collide = info->Get();
-
-	if (collide->num_hits <= 0) 
-	{
-		return ade_set_args(L, "f", -1.0f);;
-	}
-	else
-	{
-		return ade_set_args(L, "f", collide->hit_dist);
-	}
-}
-
-ADE_FUNC(getCollisionPoint, l_ColInfo, "[boolean local]", "The collision point of this information (local to the object if boolean is set to <i>true</i>)", "vector", "The collision point or nil of none")
-{
-	mc_info_h* info;
-	bool local = false;
-
-	if(!ade_get_args(L, "o|b", l_ColInfo.GetPtr(&info), &local))
-		return ADE_RETURN_NIL;
-
-	if (!info->IsValid())
-		return ADE_RETURN_NIL;
-
-	mc_info *collide = info->Get();
-	
-	if (collide->num_hits <= 0) 
-	{
-		return ADE_RETURN_NIL;
-	}
-	else
-	{
-		if (local)
-			return ade_set_args(L, "o", l_Vector.Set(collide->hit_point));
-		else
-			return ade_set_args(L, "o", l_Vector.Set(collide->hit_point_world));
-	}
-}
-
-ADE_FUNC(getCollisionNormal, l_ColInfo, "[boolean local]", "The collision normal of this information (local to object if boolean is set to <i>true</i>)", "vector", "The collision normal or nil of none")
-{
-	mc_info_h* info;
-	bool local = false;
-
-	if(!ade_get_args(L, "o|b", l_ColInfo.GetPtr(&info), &local))
-		return ADE_RETURN_NIL;
-
-	if (!info->IsValid())
-		return ADE_RETURN_NIL;
-
-	mc_info *collide = info->Get();
-
-	if (collide->num_hits <= 0) 
-	{
-		return ADE_RETURN_NIL;
-	}
-	else
-	{
-		if (!local)
-		{
-			vec3d normal;
-
-			vm_vec_unrotate(&normal, &collide->hit_normal, collide->orient);
-
-			return ade_set_args(L, "o", l_Vector.Set(normal));
-		}
-		else
-		{
-			return ade_set_args(L, "o", l_Vector.Set(collide->hit_normal));
-		}
-	}
-}
-
-ADE_FUNC(isValid, l_ColInfo, NULL, "Detects if this handle is valid", "boolean", "true if valid false otherwise")
-{
-	mc_info_h* info;
-
-	if(!ade_get_args(L, "o", l_ColInfo.GetPtr(&info)))
-		return ADE_RETURN_NIL;
-
-	if (info->IsValid())
-		return ADE_RETURN_TRUE;
-	else
-		return ADE_RETURN_FALSE;
-}
 
 //**********HANDLE: Object
 ade_obj<object_h> l_Object("object", "Object handle");
